requirements.txt
openai==1.61.1
numpy==1.26.4
python-dotenv==1.0.1  # for managing environment variables 
pytest==8.0.0  # for testing
